pyproject.toml
[build-system]
requires = ["setuptools>=70.0"]
build-backend = "setuptools.build_meta"

[project]
name = "comfyui-save-file-extended"
version = "0.0.5"
description = "A collection of custom nodes for ComfyUI that offers an extension to the existing Save and Preview Image/Video nodes allowing directly adding and previewing the files from your preferred cloud storage providers (S3/DropBox/Google Drive/...)"
authors = [
  {name = "RUiNtheExtinct", email = "deepkarma001@gmail.com"},
  {name = "Prem Sai G", email = "premsaig1605@gmail.com"},
]
readme = "README.md"
license = {text = "MIT license"}
requires-python = ">=3.10"
classifiers = []
dependencies = [
    "boto3>=1.26.0",
    "google-cloud-storage>=3.3.0",
    "azure-storage-blob>=12.19.0",
    "b2sdk>=2.10.0",
    "dropbox>=12.0.2",
    "requests>=2.31.0",
    "supabase>=2.4.0",
    "av>=14.4.0",
]


[project.optional-dependencies]
dev = [
    "bump-my-version",
    "coverage",  # testing
    "mypy",  # linting
    "pre-commit", # runs linting on commit
    "pytest",  # testing
    "pytest-cov",  # coverage plugin for pytest
    "ruff",  # linting
    # Test dependencies (also used by nodes, but provided by ComfyUI in production)
    "torch",  # for creating test data and node functionality
    "torchaudio",  # for audio node tests
    "numpy",  # for image/audio processing tests
    "pillow",  # for image processing tests
]

[project.urls]
Repository = "https://github.com/RUiNtheExtinct/comfyui-save-file-extended"
BugTracker = "https://github.com/RUiNtheExtinct/comfyui-save-file-extended/issues"
Documentation = "https://github.com/RUiNtheExtinct/comfyui-save-file-extended/wiki"


[tool.comfy]
PublisherId = "ruintheextinct"
DisplayName = "comfyui-save-file-extended"
Icon = ""
Tags = []
Repository = "https://github.com/RUiNtheExtinct/comfyui-save-file-extended"

includes = []

[tool.setuptools.package-data]
"*" = ["*.*"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = [
    "tests",
]

[tool.mypy]
files = "."

# Use strict defaults
strict = true
warn_unreachable = true
warn_no_return = true

[[tool.mypy.overrides]]
# Don't require test functions to include types
module = "tests.*"
disable_error_code = ["attr-defined", "no-untyped-def"]

[tool.ruff]
# extend-exclude = ["static", "ci/templates"]
line-length = 140
src = ["src", "tests"]
target-version = "py39"

# Add rules to ban exec/eval
[tool.ruff.lint]
select = [
    "S102",  # exec-builtin
    "S307",  # eval-used
    "W293",
    "F",  # The "F" series in Ruff stands for "Pyflakes" rules, which catch various Python syntax errors and undefined names.
    # See all rules here: https://docs.astral.sh/ruff/rules/#pyflakes-f
]
ignore = [
    "F403",  # Allow star imports (from module import *)
    "F405",  # Allow star imports (from module import *)
    'W293',  # Allow blank line contains whitespace
    "F841",  # Allow unused variable
]

[tool.ruff.lint.flake8-quotes]
inline-quotes = "double"
